Adds GET methods to CourseManagement view class.
